core/fdt: Allow to specify FDT blob
The FDT memory block is allocated by skiboot and passed to kernel
after the device tree is flattenned. In order to transfer device
tree changes for PCI hotplug in subsequent patches, new OPAL API
will be introduced to retrieve the sub-tree behind the specified
(PCI slot) node. The FDT blob is allocated by kernel in that case.

This adds parameters for various functions in core/fdt.c allowing
to specify the memory block for the FDT blob, which is allocated
by skiboot or kernel.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>