core/fdt: Allow to exclude root node
The root node is excluded in the device sub-tree created during
PCI hot add time. This adds one extra argument @exclusive to
flatten_dt_node(), __create_dtb() and create_dtb() to indicate
the root node should be excluded or not. The changes are going
to be used by PCI hot add path and it's not affecting anything
at present.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>