core/fdt: Allow to specify FDT blob
There is a static variable @fdt pointing the FDT blob. The memory
block is allocated by skiboot. After the device tree is flattened,
it's passed to kernel when it's loaded. It works fine. The FDT blob
for device sub-tree because of PCI hot add is allocated by kernel,
using a static @fdt to dereferencing the FDT blob isn't correct any
more.

This adds @fdt argument to various functions to specify the target
FDT blob memory block. As a result, the file scoped static variable
@fdt is removed.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>