core/pci: Fix wrong reserved PE# in enumeration
When scanning to non-existing PCI device, EEH (frozen) error is
usually happening. We clear the unexpected frozen PE state after
it. The reserved PE number is assumed to be 0 wrongly. So the
frozen state on the reserved PE number isn't cleared properly.

This introduces struct phb_ops::get_info() to retrieve the reserved
PE number from platforms. Then the EEH frozen state checking and
clearing are applied to the reserved PE number.

Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>