phb: Work around XSL bug sending PTE updates with wrong scope
The XSL incorrectly sends out PTE updates with nodal scope instead of
group scope, which causes a system checkstop. Work around this issue by
forcing all commands to be unlimited scope.

The performance impact should be negligible since the XSL is only used
for translations, not for block data transfers.

Signed-off-by: Ian Munsie <imunsie@au1.ibm.com>