hw/phb3: disable CAPI mode during creset
When fast rebooting or kexec-ing a system with a PHB in CAPI mode, we need
to return the PHB to regular PCIe mode.

In order to do this, we reset a bunch of registers to their pre-CAPI-mode
state.

However, disabling these registers while there is traffic going over the
PCI link is dangerous and will generally cause a checkstop.

As such, we want to do this while the PHB is fenced. Conveniently, during a
creset we force a PHB fence, so this is a good opportunity to disable CAPI
mode.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>