fast-reboot: abort fast reboot if CAPP attached
If a PHB is in CAPI mode, we cannot safely fast reboot - the PHB will be
fenced during the reboot resulting in major problems when we load the new
kernel.

In order to handle this safely, we need to disable CAPI mode before
resetting PHBs during the fast reboot. However, we don't currently support
this.

In the meantime, when fast rebooting, check if there are any PHBs with a
CAPP attached, and if so, abort the fast reboot and revert to a normal
reboot instead.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

Lightly tested. I'm still reworking my patch to disable the CAPP properly
during a PHB creset which will render this unneccessary.