hw/phb3: fix error handling in complete reset
During a complete reset, when we get a timeout waiting for pending
transaction in state PHB3_STATE_CRESET_WAIT_CQ, we mark the PHB as broken
and return OPAL_PARAMETER.

Change the return code to OPAL_HARDWARE which is way more sensible, and set
the state to PHB3_STATE_UNINITIALIZED so that the kernel can retry the
complete reset.

Reported-by: Pradipta Ghosh <pradghos@in.ibm.com>
Suggested-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

This patch applies on top of skiboot 5.1.18.

I'm going to submit a similar fix for mainline along with a few other
fixes around this part of the code, but 5.1.18 is what we're using for
Surelock testing.

See bugzilla 144323 for more details.