hmi: Print CAPP FIR information when handling CAPP malfunction alerts
When diagnosing or debugging CAPP errors, it's rather useful to have the
CAPP FIR, which often provides very helpful information.

Print the CAPP FIR to the log when we handle a Malfunction Alert HMI for a
CAPP error.

Cc: Frederic Barrat <fbarrat@linux.vnet.ibm.com>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

This is based off skiboot 5.3.6, for SW374395. Completely untested although
based on some code which has been tested a couple of times...