Initial OpenCAPI Support for Zaius Platform
This patch implements basic OpenCAPI functionality for testing purposes
on Zaius machines. Slot/brick configuration is hardcoded in platform
code to declare a link on brick 3 - this can be modified for brick 2, other
bricks are not expected to work at this time. A temporary I2C workaround
to enable card reset is also included in platform code.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>