p9-adu: implement update sequence for PowerBus Hotplug Mode Control
The PowerBus Hotplug Mode Control register is used to control various
things about the PowerBus.

Due to hardware timing constraints, updating this register requires
more than just a simple SCOM write. There are two sets of pb_hp_mode
registers, current (B) and next (A). Updating is done by a SCOM write
to the next registers, followed by a SWITCH_AB pulse from the
Alter/Display Unit, which triggers a hardware sequence to swap
between the current and next registers.

The sequence used in this code was derived from the hardware
procedures used by Hostboot.

This will be used in a subsequent patch as part of the
OpenCAPI/NVLink init sequence.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>