README: Fix/improve Ubuntu/Fedora build instructions
The current instructions for Ubuntu/Fedora are missing a couple of dependencies
and in the Ubuntu case the CROSS= prefix doesn't match the compiler package name.

Add missing xz and make dependencies. Recommend the use of a powerpc64le
bi-endian toolchain, and change dependencies and CROSS= variables to match that.

Fixes: 8fbf67e5b1bd ("README: Update build instructions")
Reported-by: Mark Linimon <linimon@freebsd.org>
Suggested-by: Jeremy Kerr <jk@ozlabs.org>
Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>

---

v1->v2
	* recommend installation of a LE/bi-endian toolchain, per JK
	* add xz dependency for both fedora and ubuntu
	* add make dependency for ubuntu