platforms: Add OpenCAPI platform data and device tree nodes
Add OpenCAPI platform data for the zaius and zz platforms, as well as the
generic platform that's used for BML boots. Currently, all three platforms
are identical, but that won't always be the case (e.g. Witherspoon).

Additionally, for Zaius, hardcode link information and the I2C bus that's
required to reset OpenCAPI devices, because this currently can't be
extracted from HDAT. Eventually this will go away once Hostboot adds the
relevant data in HDAT. A later patch will add this for ZZ once we've done
more testing.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>

---

v1->v2:

  - Remove extraneous break that stops things after the first CPU (Fred)
  - Remove i2c odl01 data (Fred)
  - Remove i2c 1.8v hack (Fred)
  - Don't try to create device tree nodes on zaius if they already exist
    (Fred)

v2->v3:

  - Reverse bit order of lane masks