hw/p8-i2c: Fix i2c request timeout
Commit eb146fac9685 ("core/i2c: Move the timeout field into
i2c_request") simplified a bit how a request timeout is
handled. However there's now some confusion between milliseconds and
timebase increments when defining or using the timeout values, which
breaks i2c requests made for opencapi, and probably others too.

This patch declares all the timeout in milliseconds and just converts
to timebase at the end of the chain, as needed.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>