hw/npu2: Replace npu2_dev::bars array with individual BARs
We currently maintain the state of the NTL and GENID BARs for each device
in npu2_dev::bars[2], with the first BAR being the NTL BAR and the second
being the GENID BAR.

Given we only have two BARs, it's a lot clearer to get rid of the array and
just have two struct members, ntl_bar and genid_bar.

No functional change.

Signed-off-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Alexey Kardashevskiy <aik@ozlabs.ru>

---
v1->v2:
- split this patch out